pyproject.toml
[tool.poetry]
name = "orkg-abstracts"
version = "0.1.0"
description = "Package for fetching ORKG paper abstracts and dumping them"
authors = ["Omar Arab Oghli"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.1"
pyoai = "2.5.0"
rdflib = "6.1.1"
pandas = "1.3.5"
tqdm = "4.64.0"
requests = "2.28.1"
SPARQLWrapper = "2.0.0"
python-dotenv = "0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
